+++ shared/recover/development/jpcsp.cmd
@@ -1,29 +1,33 @@
-@set PATH=%CD%\Git\bin;%CD%\Git\mingw64\bin;%CD%\Git\cmd;%PATH%
+@cd ..\..\..\..\..\
+@set PATH=%CD%\Git\bin;%CD%\Git\mingw64\bin;%CD%\Git\cmd;%CD%\..\Java\JDK\bin;%ProgramFiles%\Java\JDK\bin;%CD%\..\7-ZipPortable\App\64;%PATH%
 @if not exist "%HOME%" @set HOME=%HOMEDRIVE%%HOMEPATH%
 @if not exist "%HOME%" @set HOME=%USERPROFILE%
 
 @set PLINK_PROTOCOL=ssh
 @if not defined TERM set TERM=msys
 @cd projects
-@If NOT exist "ppsspp"\ (
-git clone https://github.com/hrydgard/ppsspp.git ppsspp
-@cd ppsspp
+@If NOT exist "jpcsp"\ (
+git clone https://github.com/jpcsp/jpcsp.git jpcsp
+@cd jpcsp
 git submodule update --init --recursive
 )
-@If exist "ppsspp"\ (
-@cd ppsspp
+@If exist "jpcsp"\ (
+@cd jpcsp
 )
 git branch work
 
+@set netbeanspath=..\..\..\
+@if not exist %netbeanspath%netbeans @set netbeanspath=C:\Progra~1\
+
 :Choice
-@ECHO. 
+@echo.
 @echo ----------------------- 
 @echo What do you want to do?
 @echo -----------------------
-@echo 1. Start Visual Studio
-@echo 2. Update local repository
-@echo 3. Update forked repository
-@echo 4. Update personal build
+@echo 1. Start NetBeans
+@echo 2. Build project
+@echo 3. Update local repository
+@echo 4. Update forked repository
 @echo 5. Launch GIT GUI
 @echo 6. Update submodules
 @echo 7. View repository status
@@ -33,11 +37,12 @@ git branch work
 @echo 11. Clean build and untracked files and folders
 @echo 12. Insert commands manually
 @echo 13. Exit
+
 @set /p choice="Enter your Choice here:"
-@if %choice%==1 GOTO Start_VS 
-@if %choice%==2 GOTO Update_local
-@if %choice%==3 GOTO Update_remote
-@if %choice%==4 GOTO Update_build
+@if %choice%==1 GOTO Start_NetBeans 
+@if %choice%==2 GOTO Build
+@if %choice%==3 GOTO Update_local
+@if %choice%==4 GOTO Update_remote
 @if %choice%==5 GOTO GUI
 @if %choice%==6 GOTO Update_submodules  
 @if %choice%==7 GOTO Status
@@ -47,12 +52,18 @@ git branch work
 @if %choice%==11 GOTO Clean_build
 @if %choice%==12 GOTO Command
 @if %choice%==13 GOTO Exit
-  
-:Start_VS
-@set /p update-ver="Update git version before building in Visual studio Y/N? "
-@if %update-ver%==y @START Windows\git-version-gen.cmd ^&^& exit
-@if %update-ver%==Y @START Windows\git-version-gen.cmd ^&^& exit
-@Windows\PPSSPP.sln
+
+
+:Start_NetBeans
+@%netbeanspath%netbeans\bin\netbeans64.exe
+@GOTO Choice
+
+:Build
+@copy dist\jpcsp-windows-amd64\Settings.properties .
+@RD /S /Q bin
+@RD /S /Q ms0\PSP\SAVEDATA
+@RD /S /Q dist\jpcsp-windows-amd64\ms0\PSP\SAVEDATA
+@start %netbeanspath%netbeans\extide\ant\bin\ant -f build-auto.xml dist-windows-amd64 ^&^& del dist\*.7z ^&^& mklink /J dist\jpcsp-windows-amd64\ms0\PSP\SAVEDATA ..\..\..\ppsspp\memstick\PSP\SAVEDATA ^&^& copy Settings.properties dist\jpcsp-windows-amd64 ^&^& pause ^&^& exit
 @GOTO Choice
 
 :Update_local
@@ -66,22 +77,12 @@ git rebase master
 @GOTO Choice
 
 :Update_remote
-git remote add upstream https://github.com/pal1000/ppsspp.git
+git remote add upstream https://github.com/pal1000/jpcsp.git
 git fetch upstream
-git push -f --all upstream
+@set /p push=Update your fork (y/n):
+if /I "%push%"=="y" git push -f --all upstream
 @GOTO Choice
 
-:Update_build
-RD /S /Q ..\..\..\ppsspp\assets
-RD /S /Q "..\..\..\ppsspp\memstick\PSP\PPSSPP_STATE"
-DEL "..\..\..\ppsspp\PPSSPPWindows64.exe"
-DEL "..\..\..\ppsspp\PPSSPPWindows.exe"
-DEL ..\..\..\ppsspp\memstick\PSP\SYSTEM\ppsspp.ini
-XCOPY assets ..\..\..\ppsspp\assets /S /E /I /Q
-copy "PPSSPPWindows64.exe" ..\..\..\ppsspp
-copy "PPSSPPWindows.exe" ..\..\..\ppsspp
-@GOTO Choice
- 
 :GUI
 @git gui
 @GOTO Choice
@@ -116,7 +117,7 @@ git clean -d -f -x
 @GOTO Choice
 
 :Command
-@start %COMSPEC%
+start %COMSPEC%
 @GOTO Choice
 
 :Update_submodules
@@ -125,5 +126,3 @@ git submodule update --init --recursive
 
 :Exit
 exit
-
-
